% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hit.R
\name{hit}
\alias{hit}
\title{Find expected hitting times}
\usage{
hit(mc, e = 0.01)
}
\arguments{
\item{mc}{of type mc containing a defined markov chain}

\item{e}{floating point convergence criterion for infinite mc.}
}
\value{
The submatrix of expected hitting times for mc.
}
\description{
Input an mc to get a submatrix of hitting times where M[i,j] is the E(Tij).
}
\examples{
pijdef = matrix(rep(0,9), nrow=3)
pijdef[1,1] <- 0.5
pijdef[1,2] <- 0.5
pijdef[2,3] <- 0.5
pijdef[2,1] <- 0.5
pijdef[3,1] <- 1
markov_chain = mc(pijdef)
hit(markov_chain, 0.1)
}

