% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hit.R
\name{hit}
\alias{hit}
\title{Find expected hitting times}
\usage{
hit(mc, e = 0.01)
}
\arguments{
\item{mc}{of type mc containing a defined pijdef.}

\item{e}{floating point convergence criterion for infinite pijdef.}
}
\value{
The submatrix of expected hitting times for mc$pijdef.
}
\description{
Input an mc to get a submatrix of hitting times where M[i,j] is the E(Tij).
}

