% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc.R
\name{mc}
\alias{mc}
\title{Markov chain definition.}
\usage{
mc(pijdef = NULL)
}
\arguments{
\item{Construct}{and initialize with some pijdef.}
}
\description{
Markov chain definition.
}
\section{Slots}{

\describe{
\item{\code{pijdef}}{is either a square finite transition matrix
or a function(i, j) defining a potentially infinite transition matrix.}
}}

