% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stn.R
\name{stn}
\alias{stn}
\title{Find stationary distribution}
\usage{
stn(mc, e = 0.01)
}
\arguments{
\item{mc}{of type mc containing a defined pijdef.}

\item{e}{floating point convergence criterion for infinite pijdef.}
}
\value{
The stationary distribution vector for mc$pijdef.
}
\description{
Input an mc to get a vector of stationary probabilities for each markov state.
}

