% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stn.R
\name{stn}
\alias{stn}
\title{Find stationary distribution}
\usage{
stn(mc, e = 0.01)
}
\arguments{
\item{mc}{of type mc containing a defined pijdef.}

\item{e}{floating point convergence criterion for infinite pijdef.}
}
\value{
The stationary distribution vector for mc$pijdef.
}
\description{
Input an mc to get a vector of stationary probabilities for each markov state.
}
\examples{
pijdef = matrix(rep(0,9), nrow=3)
pijdef[1,1] <- 0.5
pijdef[1,2] <- 0.5
pijdef[2,3] <- 0.5
pijdef[2,1] <- 0.5
pijdef[3,1] <- 1
markov_chain = mc(pijdef)
stn(markov_chain, 0.1)
}

